Move OfflineTaskService related files into separate folder.
