+++ src/wallabag.Data/ViewModels/AddItemViewModel.cs
@@ -6,6 +6,7 @@ using System.Windows.Input;
 using wallabag.Data.Common.Helpers;
 using wallabag.Data.Models;
 using wallabag.Data.Services;
+using wallabag.Data.Services.OfflineTaskService;
 
 namespace wallabag.Data.ViewModels
 {
