+++ src/wallabag.Data/ViewModels/MainViewModel.cs
@@ -17,6 +17,7 @@ using wallabag.Data.Common.Messages;
 using wallabag.Data.Interfaces;
 using wallabag.Data.Models;
 using wallabag.Data.Services;
+using wallabag.Data.Services.OfflineTaskService;
 
 namespace wallabag.Data.ViewModels
 {
